move load_config and retore in automl.common.util to chronos.feature.utils
