+++ pyzoo/test/zoo/chronos/feature/test_time_sequence_feature.py
@@ -17,8 +17,7 @@
 import pytest
 
 from test.zoo.pipeline.utils.test_utils import ZooTestCase
-from zoo.automl.common.util import restore
-from zoo.chronos.feature.utils import save
+from zoo.chronos.feature.utils import save, restore
 from zoo.chronos.feature.time_sequence import *
 from numpy.testing import assert_array_almost_equal
 import json
