+++ pyzoo/zoo/automl/common/util.py
@@ -20,13 +20,9 @@ import zipfile
 import os
 import json
 
+from zoo.chronos.feature.utils import restore
+
 IDENTIFIER_LEN = 27
-
-
-def load_config(file_path):
-    with open(file_path, "r") as input_file:
-        data = json.load(input_file)
-    return data
 
 
 def process(command, fail_fast=False, timeout=120):
@@ -101,22 +97,6 @@ def get_ckpt_hdfs(remote_dir, local_ckpt):
 
     cmd = "hadoop fs -get {} {}".format(remote_ckpt, local_ckpt_dir)
     process(cmd)
-
-
-def restore(file, feature_transformers=None, model=None, config=None):
-    model_path = os.path.join(file, "weights_tune.h5")
-    config_path = os.path.join(file, "config.json")
-    local_config = load_config(config_path)
-    if config is not None:
-        all_config = config.copy()
-        all_config.update(local_config)
-    else:
-        all_config = local_config
-    if model:
-        model.restore(model_path, **all_config)
-    if feature_transformers:
-        feature_transformers.restore(**all_config)
-    return all_config
 
 
 def restore_hdfs(model_path, remote_dir, feature_transformers=None, model=None, config=None):
